Wan pipeline end2end working & Add registry for scheduler & Add interface for scheduler+DiT

+++ fastvideo/v1/pipelines/stages/latent_preparation.py
@@ -9,6 +9,7 @@ from fastvideo.v1.logger import init_logger
 from fastvideo.v1.pipelines.pipeline_batch_info import ForwardBatch
 from fastvideo.v1.pipelines.stages.base import PipelineStage
 from fastvideo.v1.models.vaes.common import ParallelTiledVAE
+import torch
 
 logger = init_logger(__name__)
 
@@ -90,7 +91,7 @@ class LatentPreparationStage(PipelineStage):
             latents = randn_tensor(shape,
                                    generator=generator,
                                    device=device,
-                                   dtype=dtype)
+                                   dtype=torch.float32)
         else:
             latents = latents.to(device)
